requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
mongoengine==0.29.1
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.9.0